feat(signin): create the sign in feature -create UserController file -add login for admin, vendor and buyer -create proper error messages -included translation english and french [finishes #13]
