+++ src/database/config/config.json
@@ -1,9 +1,9 @@
 {
   "development": {
     "username": "postgres",
-    "password": "54321",
+    "password": "123",
     "database": "ecommerce-app-cogito-db-dev",
-    "host": "postgres",
+    "host": "localhost",
     "port": "5432",
     "dialect": "postgres"
   },
