+++ src/database/models/user.js
@@ -21,7 +21,8 @@ module.exports = (sequelize, DataTypes) => {
       preferredCurrency: { type: DataTypes.STRING, field: 'preferred_currency' },
       billingAddress: { type: DataTypes.ARRAY(DataTypes.STRING), field: 'billing_address' },
       password: DataTypes.STRING,
-      role: DataTypes.STRING,
+      roleId: DataTypes.INTEGER,
+      status: DataTypes.STRING,
       createdAt: { type: DataTypes.DATE, field: 'created_at' },
       updatedAt: { type: DataTypes.DATE, field: 'updated_at' },
     },
