+++ src/locales/en/translation.json
@@ -1,3 +1,4 @@
 {
     "welcome_message": "Welcome to Cogito's Ecommerce app API"
+    "loginUser_message": "Please provide email, and password to log in!
 }
